Last working stage for new transmittal.
The new transmittal gets immediately saved to the database. It is not possible to differentiate between a transmittal that is still being worked on (and further changes are allowed) and a transmittal that has been issued (and further changes are forbidden).
What's more, I do not like that I do not have a "working" VDS and I invest so much time on other parts of the functionality. I need a new branch to take care of that first (although a new branch does not make sense).