Added 'uploader' app
The goal is to make it possible to import a CSV file as starting point for the VDS. This is useful (a) as a default VDS, (b) as starting point for an active/past project.

Let's first get the CSV file into the system.
Then check validity of the CSV.
Lastly, update the VDS table(s).